requirements.txt
# Telegram Bot API library
python-telegram-bot==20.7

# Telethon for Telegram client operations
telethon==1.34.0

# Environment variable management
python-dotenv==1.0.0

# JSON handling (built-in, but listed for clarity)
# json (built-in)

# Date and time handling (built-in)
# datetime (built-in)

# Path handling (built-in)
# pathlib (built-in)

# Logging (built-in)
# logging (built-in)

# HTML escaping (built-in)
# html (built-in)

# Async support (built-in in Python 3.7+)
# asyncio (built-in)